[namespaced-rake-tasks] Eliminate building OpenStruct around names
Initially, RakeTaskManager and TaskToWebBuilder were a single object.

This means that TaskToWebBuilder (and the associated views) was likely
to "know" stuff specific to Rake tasks, and sure enough, the
RakeTaskManager knew that Rake tasks responding to a "name" method,
and this proliferated a wee bit.

This commit eliminates this. It's now a TaskManager's job to only
provide a list of named tasks (as strings) and know how to execute
those tasks when handed back a string.